32-bit sheets
Extend the 32-bit support to the new j2a-sheet.py as well. If even a single sprite is 32-bit, the entire png it's in will be rendered as 32-bit (with color 0 replaced with full transparency), though if separate rows are separate pngs, each row will be evaluated independently as 8-bit vs 32-bit.